requirements.txt
git+https://github.com/django-nonrel/django.git
Babel==2.3.4
Jinja2==2.8
MarkupSafe==0.23
Pillow==3.0.0
PyYAML==3.11
amqp==1.4.8
anyjson==0.3.3
argparse==1.2.1
billiard==3.3.0.22
blinker==1.4
celery==3.1.19
debtcollector==1.4.0
django-mongodb-engine==0.6.0
git+https://github.com/umutbozkurt/django-rest-framework-mongoengine.git
django-rest-swagger==0.3.4
djangorestframework==3.2.4
djangotoolbox==1.8.0
funcsigs==1.0.2
iso8601==0.1.11
kombu==3.0.32
mongoengine==0.10.0
monotonic==1.1
msgpack-python==0.4.7
netaddr==0.7.18
netifaces==0.10.4
oslo.config==3.9.0
oslo.context==2.3.0
oslo.i18n==3.6.0
oslo.log==3.7.0
oslo.serialization==2.6.0
oslo.utils==3.10.0
nltk==3.2.1
overpy==0.3.1
pbr==1.9.1
pyes==0.99.6
pyinotify==0.9.6
pymongo==3.0.3
python-dateutil==2.5.3
pytz==2015.7
requests==2.9.1
six==1.10.0
stevedore==1.13.0
urllib3==1.15.1
wrapt==1.10.8
wsgiref==0.1.2

####
# Additional dependencies, not covered in PIP
# (1) MONGODB  
# (2) sudo apt-get install libjpeg8-dev
# (3) sudo apt-get install rabbitmq-server

